# sycl-exp : Re-enabled mul_mat_batched_sycl Path for batched Q*K & KQ*V