This PR re-enables the MUL_MAT path for QK and QKV operations involving MKL's `gemm_batch` through `ggml_sycl_mul_mat_batched_sycl`(otherwise unused as the `src1->type() == GGML_TYPE_F16` is never realized) .
Prompt Processing performance is restored, for e.g. :

- Nvidia A100 + 70B Q4_K (b=p=512): 241 t/s -> 538 t/s
- Intel Arc A770 + 13B Q4_K (b=p=512) : 363 t/s -> 587 t/s
- Nvidia A4000 + 7B Q4_K (b=p=512) : 1710 t/s -> 2033 t/s

Text Generation performance remained the same / showed very small improvements.

- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [x] Medium
  - [ ] High