@joeatodd 2024-06-24
@airMeng this PR might restore the perf regression you describe [here](https://github.com/ggerganov/llama.cpp/discussions/5277#discussioncomment-9854608)?