@joeatodd 2024-06-24
> Outstanding to catch this small mistake!

Thanks for the feedback @airMeng  - be aware this PR is to our temporary perf branch `codeplay/sycl-main`. You will want it in `master` too I guess!